Two non-parsing tests in the parsing module
